Push multi-arch images (amd64/arm64) into docker hub.
This uses buildx to build a multi-arch image for pushing into github.

The `build` stage might not make a whole lot of sense now, but it does
build the arch that matches the host and loads it so you could use it
locally. If you invoke `docker-publish` right after it should reuse the
cache for the image for that arch and then proceed to also build the
other, pushing them into the remote registry.